+++ lib/draggable_menu.dart
@@ -2,5 +2,6 @@ library draggable_menu;
 
 export 'src/draggable_menu/menu/menu.dart';
 export 'src/draggable_menu/menu/enums/status.dart';
-export 'src/draggable_menu/menu/enums/ui.dart';
+export 'src/draggable_menu/menu/ui.dart';
 export 'src/draggable_menu/utils/scrollable_manager/scrollable_manager.dart';
+export 'src/draggable_menu/menu/custom_draggable_menu.dart';