+++ lib/src/draggable_menu/menu/ui/modern_soft.dart
@@ -1,27 +1,24 @@
+import 'package:draggable_menu/draggable_menu.dart';
 import 'package:draggable_menu/src/default/colors.dart';
 import 'package:draggable_menu/src/draggable_menu/menu/widgets/default_bar_item.dart';
 import 'package:flutter/material.dart';
 
-class SoftModernUi extends StatelessWidget {
-  final Widget? child;
+class SoftModernDraggableMenu extends CustomDraggableMenu {
   final Widget? barItem;
   final Color? accentColor;
   final Color? color;
   final double? radius;
-  final double menuValue;
 
-  const SoftModernUi({
-    super.key,
-    this.child,
+  const SoftModernDraggableMenu({
     this.barItem,
     this.accentColor,
     this.color,
     this.radius,
-    required this.menuValue,
   });
 
   @override
-  Widget build(BuildContext context) {
+  Widget buildUi(BuildContext context, Widget child, DraggableMenuStatus? status,
+      double menuValue) {
     return Padding(
       padding: EdgeInsets.all(16 * (1 - (menuValue < 0 ? 0 : menuValue))),
       child: ClipRRect(
@@ -39,7 +36,7 @@ class SoftModernUi extends StatelessWidget {
               Center(
                 child: barItem ?? DefaultBarItem(color: accentColor),
               ),
-              if (child != null) Flexible(child: child!),
+              Flexible(child: child),
             ],
           ),
         ),