+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## 0.3.0
+## 1.0.0
 
 * Created fixed Threshold parameters for the Menu Thresholds to specify them with fixed heights.